nf RDM-216 загрузка данных версии из xml-файла

See merge request microservices/rdm!109

+++ rdm-impl/src/main/java/ru/inovus/ms/rdm/service/DraftServiceImpl.java
@@ -43,8 +43,7 @@ import ru.inovus.ms.rdm.validation.PrimaryKeyUniqueValidation;
 import ru.inovus.ms.rdm.validation.ReferenceValidation;
 import ru.kirkazan.common.exception.CodifiedException;
 
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.*;
@@ -52,10 +51,10 @@ import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 
 import static java.util.Collections.emptySet;
 import static java.util.Collections.singletonList;
+import static java.util.stream.Collectors.toList;
 import static org.apache.cxf.common.util.CollectionUtils.isEmpty;
 import static ru.i_novus.platform.datastorage.temporal.enums.FieldType.STRING;
 import static ru.inovus.ms.rdm.repositiory.RefBookVersionPredicates.*;
@@ -99,6 +98,11 @@ public class DraftServiceImpl implements DraftService {
     private static final String INCOMPATIBLE_NEW_STRUCTURE_EXCEPTION_CODE = "incompatible.new.structure";
     private static final String INCOMPATIBLE_NEW_TYPE_EXCEPTION_CODE = "incompatible.new.type";
     private static final String DRAFT_ATTRIBUTE_NOT_FOUND_EXCEPTION_CODE = "draft.attribute.not.found";
+    private static final String DRAFT_NOT_FOUND_EXCEPTION_CODE = "draft.not.found";
+    private static final String REFBOOK_NOT_FOUND_EXCEPTION_CODE = "refbook.not.found";
+    private static final String REFBOOK_IS_ARCHIVED_EXCEPTION_CODE = "refbook.is.archived";
+    private static final String INVALID_VERSION_NAME_EXCEPTION_CODE = "invalid.version.name";
+    private static final String INVALID_VERSION_PERIOD_EXCEPTION_CODE = "invalid.version.period";
 
     @Autowired
     @SuppressWarnings("all")
@@ -147,11 +151,16 @@ public class DraftServiceImpl implements DraftService {
 
         try {
             Supplier<InputStream> inputStreamSupplier = () -> fileStorage.getContent(fileModel.getPath());
+
             BiConsumer<String, Structure> consumer = getSaveDraftConsumer(refBookId);
+            Consumer<Map<String, String>> draftWithPassportCreator = getDraftWithPassportCreator(refBookId);
+
             String extension = FilenameUtils.getExtension(fileModel.getName()).toUpperCase();
-            CreateDraftBufferedRowsPersister rowsProcessor = new CreateDraftBufferedRowsPersister(draftDataService, consumer);
+            RowsProcessor rowsProcessor = new CreateDraftBufferedRowsPersister(draftDataService, consumer);
+            PassportProcessor passportProcessor = new CreateDraftPassportPersister(draftWithPassportCreator);
+
             try (FilePerRowProcessor persister = FileProcessorFactory.createProcessor(extension,
-                    rowsProcessor, new PlainRowMapper())) {
+                    rowsProcessor, new PlainRowMapper(), passportProcessor)) {
                 persister.process(inputStreamSupplier);
             } catch (IOException e) {
                 throw new RdmException(e);
@@ -169,19 +178,47 @@ public class DraftServiceImpl implements DraftService {
             RefBookVersionEntity lastRefBookVersion = getLastRefBookVersion(refBookId);
             RefBookVersionEntity draftVersion = getDraftByRefbook(refBookId);
             if (draftVersion == null && lastRefBookVersion == null) {
-                throw new NotFoundException(new Message("refbook.not.found", refBookId));
+                throw new NotFoundException(new Message(REFBOOK_NOT_FOUND_EXCEPTION_CODE, refBookId));
             }
-            if (draftVersion != null) {
+//            structure == null means that draft was created during passport saving
+            if (draftVersion != null && draftVersion.getStructure() != null) {
                 dropDataService.drop(Collections.singleton(draftVersion.getStorageCode()));
                 versionRepository.delete(draftVersion.getId());
-                draftVersion = newDraftVersion(structure, draftVersion);
+                draftVersion = newDraftVersion(structure, draftVersion.getPassportValues());
+            } else if (draftVersion == null) {
+                draftVersion = newDraftVersion(structure, lastRefBookVersion.getPassportValues());
             } else {
-                draftVersion = newDraftVersion(structure, lastRefBookVersion);
+                draftVersion.setStructure(structure);
             }
             RefBookEntity refBookEntity = new RefBookEntity();
             refBookEntity.setId(refBookId);
             draftVersion.setRefBook(refBookEntity);
             draftVersion.setStorageCode(storageCode);
+            versionRepository.save(draftVersion);
+        };
+    }
+
+//    creates a draft version to save passport values from file
+    private Consumer<Map<String, String>> getDraftWithPassportCreator(Integer refBookId) {
+        return passport -> {
+            RefBookVersionEntity draftVersion = getDraftByRefbook(refBookId);
+            if (draftVersion == null && getLastRefBookVersion(refBookId) == null)
+                throw new NotFoundException(new Message(REFBOOK_NOT_FOUND_EXCEPTION_CODE, refBookId));
+
+            if (draftVersion != null) {
+                dropDataService.drop(Collections.singleton(draftVersion.getStorageCode()));
+                versionRepository.delete(draftVersion.getId());
+            }
+            draftVersion = newDraftVersion(null, passport
+                    .entrySet()
+                    .stream()
+                    .map(entry -> new PassportValueEntity(new PassportAttributeEntity(entry.getKey()), entry.getValue(), null))
+                    .collect(toList()));
+
+            RefBookEntity refBookEntity = new RefBookEntity();
+            refBookEntity.setId(refBookId);
+            draftVersion.setRefBook(refBookEntity);
+
             versionRepository.save(draftVersion);
         };
     }
@@ -201,7 +238,7 @@ public class DraftServiceImpl implements DraftService {
         }
         List<Field> fields = fields(structure);
         if (draftVersion == null) {
-            draftVersion = newDraftVersion(structure, lastRefBookVersion);
+            draftVersion = newDraftVersion(structure, lastRefBookVersion.getPassportValues());
             RefBookEntity refBookEntity = new RefBookEntity();
             refBookEntity.setId(refBookId);
             draftVersion.setRefBook(refBookEntity);
@@ -226,13 +263,12 @@ public class DraftServiceImpl implements DraftService {
         draftVersion.setStructure(structure);
     }
 
-    private RefBookVersionEntity newDraftVersion(Structure structure, RefBookVersionEntity original) {
-        RefBookVersionEntity draftVersion;
-        draftVersion = new RefBookVersionEntity();
+    private RefBookVersionEntity newDraftVersion(Structure structure, List<PassportValueEntity> originalPassport) {
+        RefBookVersionEntity draftVersion = new RefBookVersionEntity();
         draftVersion.setStatus(RefBookVersionStatus.DRAFT);
-        draftVersion.setPassportValues(original.getPassportValues().stream()
+        draftVersion.setPassportValues(originalPassport.stream()
                 .map(v -> new PassportValueEntity(v.getAttribute(), v.getValue(), draftVersion))
-                .collect(Collectors.toSet()));
+                .collect(toList()));
         draftVersion.setStructure(structure);
         return draftVersion;
     }
@@ -310,17 +346,23 @@ public class DraftServiceImpl implements DraftService {
             Supplier<InputStream> inputStreamSupplier = () -> fileStorage.getContent(fileModel.getPath());
 
             StructureRowMapper nonStrictOnTypeRowMapper = new NonStrictOnTypeRowMapper(structure, versionRepository);
-            try (FilePerRowProcessor validator = FileProcessorFactory.createProcessor(extension,
-                    new RowsValidatorImpl(versionService, searchDataService, structure, storageCode, errorCountLimit,
-                            attributeValidationRepository.findAllByVersionId(draftId)), nonStrictOnTypeRowMapper)) {
+            try (FilePerRowProcessor validator = FileProcessorFactory
+                    .createProcessor(extension,
+                            new RowsValidatorImpl(versionService, searchDataService, structure, storageCode, errorCountLimit,
+                                    attributeValidationRepository.findAllByVersionId(draftId)),
+                            nonStrictOnTypeRowMapper,
+                            new PassportValidatorImpl())) {
                 validator.process(inputStreamSupplier);
             } catch (IOException e) {
                 throw new RdmException(e);
             }
 
             StructureRowMapper structureRowMapper = new StructureRowMapper(structure, versionRepository);
-            try (FilePerRowProcessor persister = FileProcessorFactory.createProcessor(extension,
-                    new BufferedRowsPersister(draftDataService, storageCode, structure), structureRowMapper)) {
+            try (FilePerRowProcessor persister = FileProcessorFactory
+                    .createProcessor(extension,
+                            new BufferedRowsPersister(draftDataService, storageCode, structure),
+                            structureRowMapper,
+                            new PassportPersister(versionService, draftId))) {
                 persister.process(inputStreamSupplier);
             } catch (IOException e) {
                 throw new RdmException(e);
@@ -360,11 +402,11 @@ public class DraftServiceImpl implements DraftService {
             if (versionName == null) {
                 versionName = versionNumberStrategy.next(refBookId);
             } else if (!versionNumberStrategy.check(versionName, refBookId)) {
-                throw new UserException(new Message("invalid.version.name", versionName));
+                throw new UserException(new Message(INVALID_VERSION_NAME_EXCEPTION_CODE, versionName));
             }
 
             if (fromDate == null) fromDate = LocalDateTime.now();
-            if (toDate != null && fromDate.isAfter(toDate)) throw new UserException("invalid.version.period");
+            if (toDate != null && fromDate.isAfter(toDate)) throw new UserException(INVALID_VERSION_PERIOD_EXCEPTION_CODE);
 
             versionPeriodPublishValidation.validate(fromDate, toDate, refBookId);
 
@@ -397,7 +439,7 @@ public class DraftServiceImpl implements DraftService {
             dropDataService.drop(dataStorageToDelete);
 
             RefBookVersion versionModel = versionService.getById(draftId);
-            for (FileType fileType : PerRowFileGeneratorFactory.getAvalibleTypes())
+            for (FileType fileType : PerRowFileGeneratorFactory.getAvailableTypes())
                 saveVersionFile(versionModel, fileType, generateVersionFile(versionModel, fileType));
         } finally {
             refBookLockService.deleteRefBookAction(refBookId);
@@ -610,7 +652,7 @@ public class DraftServiceImpl implements DraftService {
 
     private void validatePrimaryKeyUnique(String storageCode, UpdateAttribute updateAttribute) {
         List<Message> pkValidationMessages = new PrimaryKeyUniqueValidation(draftDataService, storageCode,
-                Collections.singletonList(updateAttribute.getCode())).validate();
+                singletonList(updateAttribute.getCode())).validate();
         if (pkValidationMessages != null && !pkValidationMessages.isEmpty())
             throw new UserException(pkValidationMessages);
     }
@@ -660,7 +702,6 @@ public class DraftServiceImpl implements DraftService {
 
         draftDataService.deleteField(draftEntity.getStorageCode(), attributeCode);
 
-
         attributeValidationRepository.delete(
                 attributeValidationRepository.findAllByVersionIdAndAttribute(draftId, attributeCode));
     }
@@ -708,7 +749,7 @@ public class DraftServiceImpl implements DraftService {
             validateAttributeExists(draftId, attribute);
             validations = attributeValidationRepository.findAllByVersionIdAndAttribute(draftId, attribute);
         }
-        return validations.stream().map(this::attributeValidationModel).collect(Collectors.toList());
+        return validations.stream().map(this::attributeValidationModel).collect(toList());
     }
 
     @Override
@@ -719,7 +760,7 @@ public class DraftServiceImpl implements DraftService {
         RefBookVersionEntity versionEntity = versionRepository.findOne(versionId);
         List<AttributeValidationEntity> validationEntities = validations.stream()
                 .map(validation -> new AttributeValidationEntity(versionEntity, attribute, validation.getType(),
-                        validation.valuesToString())).collect(Collectors.toList());
+                        validation.valuesToString())).collect(toList());
         validateDataBase(versionEntity, validationEntities);
 
         deleteAttributeValidation(versionId, attribute, null);
@@ -749,7 +790,7 @@ public class DraftServiceImpl implements DraftService {
 
         return new ExportFile(
                 generateVersionFile(versionModel, fileType),
-                fileNameGenerator.generateZipName(versionModel, FileType.XLSX));
+                fileNameGenerator.generateZipName(versionModel, fileType));
     }
 
     private AttributeValidation attributeValidationModel(AttributeValidationEntity validationEntity) {
@@ -761,9 +802,9 @@ public class DraftServiceImpl implements DraftService {
     }
 
     private InputStream generateVersionFile(RefBookVersion versionModel, FileType fileType) {
-        VersionDataIterator dataIterator = new VersionDataIterator(versionService, Collections.singletonList(versionModel.getId()));
+        VersionDataIterator dataIterator = new VersionDataIterator(versionService, singletonList(versionModel.getId()));
         try (FileGenerator fileGenerator = PerRowFileGeneratorFactory
-                .getFileGenerator(dataIterator, versionService.getStructure(versionModel.getId()), fileType);
+                .getFileGenerator(dataIterator, versionModel, fileType);
              Archiver archiver = new Archiver()) {
             if (includePassport) {
                 try (FileGenerator passportPdfFileGenerator = new PassportPdfFileGenerator(passportValueRepository, versionModel.getId(), passportFileHead)) {
@@ -792,25 +833,25 @@ public class DraftServiceImpl implements DraftService {
 
     private void validateRefBookNotArchived(Integer refBookId) {
         if (refBookId != null && versionRepository.exists(isVersionOfRefBook(refBookId).and(isArchived()))) {
-            throw new UserException("refbook.is.archived");
+            throw new UserException(REFBOOK_IS_ARCHIVED_EXCEPTION_CODE);
         }
     }
 
     private void validateDraftNotArchived(Integer draftId) {
         if (draftId != null && versionRepository.exists(hasVersionId(draftId).and(isArchived()))) {
-            throw new UserException("refbook.is.archived");
+            throw new UserException(REFBOOK_IS_ARCHIVED_EXCEPTION_CODE);
         }
     }
 
     private void validateRefBookExists(Integer refBookId) {
         if (refBookId == null || !versionRepository.exists(isVersionOfRefBook(refBookId))) {
-            throw new NotFoundException("refbook.not.found");
+            throw new NotFoundException(REFBOOK_NOT_FOUND_EXCEPTION_CODE);
         }
     }
 
     private void validateDraftExists(Integer draftId) {
         if (draftId == null || !versionRepository.exists(hasVersionId(draftId).and(isDraft()))) {
-            throw new NotFoundException(new Message("draft.not.found", draftId));
+            throw new NotFoundException(new Message(DRAFT_NOT_FOUND_EXCEPTION_CODE, draftId));
         }
     }
